#10 - Moved operator puzzle js to separate folder
